requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==1.1.2
future==0.18.2
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
line-bot-sdk==1.19.0
MarkupSafe==2.0.1
mecab-python3==0.996.5
mysql-connector-python==8.0.23
protobuf==3.17.3
requests==2.25.1
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
